refactor(corpus): fold features.py into feature_projection
Merge the shared dataclasses (Trade, WalletState, MarketState,
MarketMetadata, FeatureRow), pure helpers (apply_*_to_state,
compute_features, empty_*_state), HistoryProvider Protocol, and
StreamingHistoryProvider into feature_projection.py. This removes a
circular import (compute_features had a deferred import of project_row)
and consolidates the projection contract into one module.

The daemon-side LiveHistoryProvider is being removed in a follow-up
commit; only the corpus build-features path consumes these types now.

Rename the test files to match: test_features_* -> test_feature_projection_*.
