+++ src/pscanner/testing/__init__.py
@@ -1,6 +1,1 @@
-"""Helpers for tests and standalone benchmarking scripts.
-
-Shipped with the package (under ``src/pscanner/``, not ``tests/``) so
-non-test consumers — notably ``scripts/profile_live_history.py`` — can
-import them without reaching into ``tests/`` via ``sys.path`` hacks.
-"""
+"""Helpers shared between the test suite and benchmark scripts."""
